requirements.txt
# LLM Framework
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langgraph>=0.2.0

# Model Providers
ollama>=0.4.0
google-generativeai>=0.8.0

# Tools
duckduckgo-search>=6.0.0

# Configuration
python-dotenv>=1.0.0

# Web Framework (RAG App)
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9

# Vector Store & Embeddings
sentence-transformers>=2.3.0
faiss-cpu>=1.7.4

# PDF Processing
pymupdf>=1.23.0

# Document Processing
python-docx>=1.0.0

# Multi-Agent Framework
langgraph>=0.0.10
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=0.0.1
langchain-core>=0.1.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Code Quality
flake8>=7.0.0
